requirements.txt
python-telegram-bot==13.7
schedule==1.2.0
requests==2.31.0
